Move playground and andminconsole to views
